Remove QEMU setup and multi-platform building from Docker workflow
The changes in this commit are focussed on optimizing Docker's workflow. The QEMU setup step was removed as it was unnecessary for the Docker build process. Also, the Docker build process was previously running for multiple platforms, which was not required and thus has been stripped back to run on a single platform reducing the complexity and potentially the runtime as well.