Enable Docker build in GitHub workflow and optimize command order
Uncommented the line that makes the 'build-test-singularity' job depend on 'build-test-docker' to enable the Docker build in the GitHub workflow. The 'sleep' command in the 'run bencherclient' run block has also been removed to optimize the sequence of command execution, allowing services in the docker container to initialize properly before the benchmarks are run.