Add detailed Docker and Apptainer instructions to README
Expanded the README to include detailed instructions and examples on how to utilize the Docker and Apptainer/Singularity containers for running benchmarks. This addition will better guide users on the correct usage and commands for these containers. Code is also optimized to relocate the import requests within the download function to improve execution order.