Refactor code for thread-safety and data initialization
The commit redefines how executable files are downloaded and tied to machine architectures in the mopta08 benchmark. Also, modifications have been made to how data is downloaded in MaxSAT and SVM benchmarks.
Thread locks were introduced to ensure thread-safety during data download. For SVM benchmark, a method was created for data initialization such that it will only occur if the data has not been previously initialized.